requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
websockets==12.*
pydantic==2.*
nats-py==2.*